Refactoring overdue payment job
